requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
langchain==0.1.5
openai==1.6.1
python-dotenv==1.0.0
pyyaml
requests
rich